started refactor of raycaster

+++ java/src/C22501743/Ray.java
@@ -1,10 +1,10 @@
 package C22501743;
 
 public class Ray {
-    public double angle;
-    public double distance;
+    public float angle;
+    public float distance;
     public boolean vertical;
-    public Ray(double ang, double dist, boolean vert) {
+    public Ray(float ang, float dist, boolean vert) {
         angle = ang;
         distance = dist;
         vertical = vert;
